bevy_tasks

Module cfg

Configuration information for this crate.

Macros

async_executor

Indicates async_executor is used as the future execution backend.

async_io

Indicates async-io will be used for the implementation of block_on.

conditional_send

Indicates the current target requires additional Send bounds.

futures_lite

Indicates futures-lite will be used for the implementation of block_on.

multi_threaded

Indicates multithreading support.