bevy_tasks

Module futures

Utilities for working with Futures.

Functions

check_ready

Polls a future once, and returns the output if ready or returns None if it wasn’t ready yet.

now_or_never

Consumes a future, polls it once, and immediately returns the output or returns None if it wasn’t ready yet.