Struct bevy_tasks::AsyncComputeTaskPool

pub struct AsyncComputeTaskPool(/* private fields */);

A newtype for a task pool for CPU-intensive work that may span across multiple frames

See TaskPool documentation for details on Bevy tasks. Use ComputeTaskPool if the work must be complete before advancing to the next frame.

Implementations

impl AsyncComputeTaskPool

pub fn get_or_init(f: impl FnOnce() -> TaskPool) -> &'static Self

Gets the global AsyncComputeTaskPool instance, or initializes it with f.

pub fn try_get() -> Option<&'static Self>

Attempts to get the global AsyncComputeTaskPool instance, or returns None if it is not initialized.

pub fn get() -> &'static Self

Gets the global AsyncComputeTaskPool instance.

Panics

Panics if the global instance has not been initialized yet.