bevy_tasks
Struct IoTaskPool

pub struct IoTaskPool(/* private fields */);

A newtype for a task pool for IO-intensive work (i.e. tasks that spend very little time in a โwokenโ state)

See TaskPool documentation for details on Bevy tasks.

Implementations

impl IoTaskPool

pub fn get_or_init(f: impl FnOnce() -> TaskPool) -> &'static Self

Gets the global IoTaskPool instance, or initializes it with f.

pub fn try_get() -> Option<&'static Self>

Attempts to get the global IoTaskPool instance, or returns None if it is not initialized.

pub fn get() -> &'static Self

Gets the global IoTaskPool instance.

Panics

Panics if the global instance has not been initialized yet.